Add BaseLivewireWidget alongside BaseBladeWidget for dual-render widgets
Splits the widget rendering tech from widget identity so a downstream
package can swap any widget's frontend render path from Blade to Livewire
without redeclaring form schema, defaults, lifecycle hooks, or preview.

- Extract HasWidgetDefaults trait holding all static metadata defaults
  shared between Blade and Livewire bases.
- Rename current BaseWidget body to BaseBladeWidget; keep BaseWidget as
  a back-compat shim so existing widgets and stubs are untouched.
- Add BaseLivewireWidget that mounts a Livewire component via
  <livewire:dynamic-component>, passing $data as a prop and pre-rendered
  children as the default slot. livewire/livewire stays a soft dependency
  (composer suggest).
- Widen render() return type on BaseView to View|Htmlable|string so both
  bases satisfy the contract; the call site already stringifies output.
- Add BaseView::renderChildrenToHtml() for the Livewire slot path.
- Switch type checks in RegistersWidgets, LayupContent, and
  WidgetDefaultCompletenessTest from BaseWidget class to the Widget
  interface so Livewire widgets are recognised everywhere.
- Update LayupAssertions to coerce View|Htmlable|string render results.
- Demonstrate the identity-trait pattern on Heading, Button,
  NumberCounter, and Newsletter; remaining widgets unchanged.
- Add docs/advanced/livewire-widgets.md describing when to use which
  base, the slot-based children model, and how a downstream package
  re-registers a type to swap render tech.

All 1130 tests pass.
